remove "find" and "remove" methods from parent class Table

+++ include/row.h
@@ -3,11 +3,11 @@
 #include "polinom.h"
 
 struct Row {
-    unsigned int key;
+    unsigned long key;
     Polinom *polinom;
     bool is_deleted;
 
-    Row(unsigned int key, Polinom *polinom) :
+    Row(unsigned long key, Polinom *polinom) :
             key(key),
             polinom(polinom),
             is_deleted(false) {};
